fix(web): use Redis exception handler on API
This adds the Redis exception handler on the FastAPI API app.
